// src/app/types/koreaRnDModels.ts
export interface KoreaRnDModel{
    model_suffix: string;
    suffix: string;
    part1: string;
    part2: string;
    part3: string;
    part4: string;
    c_bom_model_suffix: string;
    bom_suffix: string;
    bom_part1: string;
    bom_part2: string;
    bom_part3: string;
    bom_part4: string;
}